requirements.txt
matplotlib==3.4.1
opencv_python==4.5.1.48
emoji_translate==0.1.0
numpy==1.20.2
Pillow==8.2.0
